build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@35469 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR3/DBGFReg.cpp
@@ -109,7 +109,7 @@ typedef struct DBGFREGSET
     /** The register descriptors. */
     PCDBGFREGDESC           paDescs;
     /** The number of register descriptors. */
-    size_t                  cDescs;
+    uint32_t                cDescs;
 
     /** Array of lookup records. */
     struct DBGFREGLOOKUP   *paLookupRecs;
